requirements.txt

Flask==3.0.3
line-bot-sdk==3.7.0
gunicorn==23.0.0
